image

+++ front/src/app/meds/med.component.ts
@@ -1,17 +1,18 @@
-import { Component, effect, inject, input } from '@angular/core';
+import { Component, Pipe, effect, inject, input } from '@angular/core';
 import { MedService } from './med.service';
-import { Med } from './medTypes';
+
 import { MatCardModule } from '@angular/material/card';
-import { NgClass, NgStyle } from '@angular/common';
+import { DatePipe, NgClass, NgStyle } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { AuthService } from '../auth/auth.service';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { Medication } from './medTypes';
 
 @Component({
   selector: 'app-med',
   standalone: true,
-  imports: [MatCardModule, NgStyle, MatIconModule, NgClass],
+  imports: [MatCardModule, NgStyle, MatIconModule, NgClass, DatePipe],
   template: `
     <div class="card-container">
       <mat-card class="example-card">
@@ -22,37 +23,68 @@ import { ToastrService } from 'ngx-toastr';
 
           <mat-card-title-group>
             <mat-card-title>{{ med.name }}</mat-card-title>
-            <mat-card-subtitle>{{ med.generic_name }}</mat-card-subtitle>
-            <img
-              mat-card-sm-image
-              src="https://material.angular.io/assets/img/examples/shiba2.jpg"
-            />
+            <mat-card-subtitle>{{ med.generic_name }}</mat-card-subtitle
+            ><br />
+            <img mat-card-sm-image src="http://localhost:3000/medications/images/{{med.image?._id}}" />
+
+            
           </mat-card-title-group>
         </mat-card-header>
-        <mat-card-content>
-          <span [ngStyle]="{ 'font-weight': 'bold' }"> Availability: </span
-          >{{ med.availability }}<br />
-          <span [ngStyle]="{ 'font-weight': 'bold' }">Medication Class: </span>
-          {{ med.medication_class }}
-          <p></p>
+        <div class="contentContainer">
+          <mat-card-content>
+            <span [ngStyle]="{ 'font-weight': 'bold' }"> Availability: </span
+            >{{ med.availability }}<br />
+            <span [ngStyle]="{ 'font-weight': 'bold' }"
+              >Medication Class:
+            </span>
+            {{ med.medication_class }}
 
-          @if(auth.is_logged_in()){
-          <div>
-            <button mat-button color="primary" (click)="onReview(med._id)">
-              Review</button
-            > 
+            <p></p>
 
+            @if(auth.is_logged_in()&&(auth.state$().fullname===med.added_by.fullname)){
             <button mat-button color="accent" (click)="onEdit(med._id)">
               Edit</button
-            > 
+            >&nbsp;
 
             <button mat-button color="warn" (click)="onDelete(med._id)">
-              Delete</button
-            ><br /> 
-          </div>
-          }
-          <p>Reviews</p>
-        </mat-card-content>
+              Delete
+            </button>
+
+            }
+            @if(auth.is_logged_in()&&(auth.state$().fullname!==med.added_by.fullname)){
+
+            <button mat-button color="primary" (click)="onReview(med._id)">
+              Review
+            </button>
+
+            }
+            <p>Reviews</p>
+
+            @for(review of med.reviews; track review){
+            <div clss="review">
+              <mat-card-header>
+                <mat-card-title>{{ review.by.fullname }}</mat-card-title>
+              </mat-card-header>
+              <mat-card-content>
+                <p>rating: {{ review.rating }}</p>
+                <p>Review: {{ review.review }}</p>
+                <p>Date: {{ review.date | date }}</p>
+              </mat-card-content>
+              <mat-card-actions>
+                @if(auth.is_logged_in()&& (review.by.fullname==
+                auth.state$().fullname)){
+                <button (click)="onEditReview(review._id)">Edit</button>&nbsp;
+                <button (click)="onEditReview(review._id)">Delete</button>
+
+                }@if(auth.is_logged_in()&&
+                (review.by.fullname!==auth.state$().fullname)){
+                <button>Helpful</button> &nbsp; <button>Like</button>&nbsp; }
+              </mat-card-actions>
+            </div>
+
+            }
+          </mat-card-content>
+        </div>
       </mat-card>
     </div>
   `,
@@ -61,12 +93,19 @@ import { ToastrService } from 'ngx-toastr';
     display:flex;
     justify-content:center;
     align-items:center;
+  
  
   }
   .example-card {
-  max-width: 350px;
+  width: 100%;
  margin-top:30px
   }
+   
+    .contentContainer{
+     display:flex;
+    justify-content:center;
+    align-items:center;
+    }
   
   `,
 })
@@ -75,18 +114,30 @@ export class MedComponent {
   readonly auth = inject(AuthService);
   _id = input<string>('');
   router = inject(Router);
-  med: Med = {
+  med: Medication = {
     _id: '',
     name: '',
+    first_letter: '',
     generic_name: '',
     medication_class: '',
     availability: '',
+    image: { filename: '', originalname: '' },
+    added_by: { user_id: '', fullname: '', email: '' },
+    reviews: [
+      {
+        review: '',
+        rating: 0,
+        by: { user_id: '', fullname: '' },
+        date: 0,
+      },
+    ],
   };
   #notification = inject(ToastrService);
   constructor() {
     effect(() => {
       if (this._id() !== '') {
         this.#medService.getMedById(this._id()).subscribe((response) => {
+          console.log(response.data.image);
           this.med = response.data;
         });
       }
@@ -99,11 +150,7 @@ export class MedComponent {
       this.#notification.error('something went wrong');
     }
   }
-  onReview(_id: any) {
-    if (_id) {
-      this.router.navigate(['', 'medications', 'reviews', 'add', _id]);
-    }
-  }
+
   onDelete(medication_id: any) {
     const confirmation: any = confirm('are you sure ');
     if (medication_id && confirmation) {
@@ -118,5 +165,14 @@ export class MedComponent {
         }
       });
     }
+  }
+  onReview(_id: any) {
+    if (_id) {
+      this.router.navigate(['', 'medications', 'reviews', 'add', _id]);
+    }
+  }
+  onEditReview(_id: any) {
+    //because Id is optional
+    this.router.navigate(['', 'medications', 'reviews', 'update', _id]);
   }
 }
